feat(chimera): test using chimera localization_group functionality to localize files.
The way this is expected to work is: precondition returns dict-wrapped list of paths, which is merged to job_param/context/runconfig as preconditions are executed. After they all execute, chimera will copy this entry to runconfig using matching keys in the dict and runconfig. Afterwards, chimera will localize the files using the `localize_groups` pge config feature.

Refs #15