Added the real implementation for precondition_functions.get_disp_s1_troposphere_files()
The function now gathers the lists of input CSLC and Compressed CSLC products for the job and uses them to derive the corresponding set of ECMWF files in S3.
If any of the expected ECMWF files are missing from S3, this function will raise an error to terminate the job.
However, if the STRICT_ANCILLARY_USAGE flag is set to False in settings.yaml, this function will instead return an empty list, essentially disabling the troposphere correction in the DISP-S1 SAS.