Added the check_for_empty_orbit_files() method to AsfDaacSlcDownload
This method will iterate over all downloaded orbit files to ensure they have non-zero size, otherwise a RuntimeError is raised.
This is to ensure we fail during the SLC download job, rather than waiting until the SCIFLO job to detect that an empty orbit file was downloaded.
This will make it easier to retry said downloads once a valid orbit file is available.